Add `create` command
Just works. Bad at a number of places. The command config
is hideous (part of the setup is left to the client). This should be
resolved through a setup method instead of having a command hang out
in the open, then it takes up dependencies and returns the commands
for the app.